Implementirana korpa za kupovinu; deo izlazne fakture"

+++ adriabike/src/main/java/com/ftn/adriabike/model/StavkaFakture.java
@@ -24,7 +24,7 @@ public class StavkaFakture {
 
     private Double osnovica;
 
-    private Double procenatPDV;
+    private Integer procenatPDV;
 
     private Double iznosPDV;
